requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0
Pillow>=10.0.0

# HTTP clients
aiohttp==3.9.0
requests==2.31.0
urllib3>=2.0.0

# Time zone handling
pytz==2024.1

# Utilities
feedparser==6.0.10

# Development/Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
